0. The basics of async

+++ 0x01-python_async_function/0-basic_async_syntax.py
@@ -8,7 +8,7 @@ import asyncio
 import random
 
 
-async def wait_random(max_delay: float = 10) -> float:
+async def wait_random(max_delay: int = 10) -> float:
     """
     Asynchronous coroutine that waits for a random delay between 0 and max_delay (inclusive) seconds.
